Combine utils into directory and split into files
