+++ mypy.ini
@@ -36,7 +36,7 @@ files = src/sentry/api/bases/external_actor.py,
         src/sentry/integrations/slack/message_builder/**/*.py,
         src/sentry/integrations/slack/requests/*.py,
         src/sentry/integrations/slack/unfurl/*.py,
-        src/sentry/integrations/slack/util/*.py,
+        src/sentry/integrations/slack/utils/*.py,
         src/sentry/integrations/slack/views/*.py,
         src/sentry/integrations/vsts/**/*.py,
         src/sentry/killswitches.py,
