+++ src/sentry/integrations/slack/integration.py
@@ -21,7 +21,7 @@ from sentry.utils.http import absolute_uri
 from sentry.utils.json import JSONData
 
 from .client import SlackClient
-from .utils import get_integration_type, logger
+from .utils import logger
 
 Channel = namedtuple("Channel", ["name", "id"])
 
@@ -67,7 +67,12 @@ metadata = IntegrationMetadata(
 
 class SlackIntegration(IntegrationInstallation):  # type: ignore
     def get_config_data(self) -> Mapping[str, str]:
-        return {"installationType": get_integration_type(self.model)}
+        metadata_ = self.model.metadata
+        # Classic bots had a user_access_token in the metadata.
+        default_installation = (
+            "classic_bot" if "user_access_token" in metadata_ else "workspace_app"
+        )
+        return {"installationType": metadata_.get("installation_type", default_installation)}
 
     def uninstall(self) -> None:
         """
