+++ src/sentry/integrations/slack/utils/auth.py
