+++ src/sentry/integrations/slack/views/unlink_identity.py
@@ -12,9 +12,9 @@ from sentry.web.decorators import transaction_start
 from sentry.web.frontend.base import BaseView
 from sentry.web.helpers import render_to_response
 
-from ..utils import logger
+from ..utils import logger, send_slack_response
 from . import build_linking_url as base_build_linking_url
-from . import never_cache, send_slack_response
+from . import never_cache
 
 SUCCESS_UNLINKED_MESSAGE = "Your Slack identity has been unlinked from your Sentry account."
